+++ CHANGELOG.md
@@ -5,6 +5,15 @@ Tutte le modifiche significative a questo progetto sono documentate in questo fi
 Il formato segue le linee guida di [Keep a Changelog](https://keepachangelog.com/it-IT/1.0.0/)
 e adotta il [versionamento semantico](https://semver.org/lang/it/).
 
+## [1.7.0] - 2025-06-07
+### Added
+- Implementato il principio di sussunzione nel file `SubsumptionPrinciple.java`,
+  per eliminare clausole che sono sovrainsieme di altre. È applicabile sia alle formule CNF sia alle formule E-CNF, quest'ultime ricavate tramite la trasformazione di Tseitin.
+
+### Changed
+- Rinominato il package `optimization` in `optionalfeatures`, perché il nome è più affine a ciò che necessita il progetto.
+
+
 ## [1.6.1] - 2025-06-07
 ### Fixed
 - Migliorato il contenuto dei file `.result` generati dal risolutore: